Tests and slight refactors
