+++ .github/workflows/labeler.yml
@@ -15,8 +15,8 @@ jobs:
       uses: actions/github-script@v7
       with:
         script: |
-          const { update_labels } = await import('${{ github.workspace }}/tools/pull_request_hooks/auto_label.js');
-          const new_labels = await update_labels({ github, context });
+          const { update_labels } = await import('${{ github.workspace }}/tools/pull_request_hooks/autoLabel.js');
+          const new_labels = await get_updated_label_set({ github, context });
           github.rest.issues.setLabels({
             issue_number: context.issue.number,
             owner: context.repo.owner,
